backend/requirements.txt
aiofiles==0.7.0
alembic==1.6.5
autopep8==1.5.7
awscli==1.19.87
backcall==0.2.0
boto3==1.17.87
bcrypt
botocore==1.20.87
cffi==1.14.5
click==7.1.2
colorama==0.4.3
cryptography==3.4.7
cycler==0.10.0
decorator==5.0.9
docutils==0.15.2
ecdsa==0.14.1
fastapi==0.65.1
greenlet==1.1.0
h11==0.12.0
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==0.10.0
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
numpy==1.20.3
pandas==1.2.4
parso==0.8.2
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
rsa==4.7.2
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
SQLAlchemy==1.4.15
starlette==0.14.2
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.13.4
wcwidth==0.2.5
